+++ packages/imperative/__tests__/src/example_clis/with_profiles/ProfileExampleConfiguration.ts
@@ -112,12 +112,6 @@ const config: IImperativeConfig = {
                     options: "--animal doggy",
                     description: "Create a profile-a profile with a doggy as the animal"
                 }
-            ],
-            updateProfileExamples: [
-                {
-                    options: "--animal froggy",
-                    description: "Update a profile-a profile to use froggy as the animal"
-                }
             ]
         },
         {